Move block editing methods to utils file
